lesson02 part3: divide source code into two c files and one header
Now the project contains two c-files and one h-file now. main.c contains the main() function. Other one contains
all other code.
