+++ 01-Git_Bash_Make/guess_a_number/lib.h
@@ -1,0 +1,18 @@
+/**
+ * @file lib.h
+ * library declaration.
+ * @author George Molchanov
+ * @date 2019.03.28
+ */
+
+#ifndef __lib_h__
+#define __lib_h__
+
+/**
+ * Compares the entered number with a randomly generated number.
+ * @param num entered number
+ * @return 1 if the num equals the generated number; 0 - otherwise
+ */
+_Bool chek_num(int num);
+
+#endif
